Fix dependencies in tests
