+++ crates/compiler/src/parser.rs
@@ -1,8 +1,8 @@
 //! The parser for the compiler.
 
 mod actual_types;
-pub mod generated;
+pub(crate) mod generated;
 mod indent_aware_lexer;
 
-pub use actual_types::*;
-pub use indent_aware_lexer::IndentAwareYarnSpinnerLexer as YarnSpinnerLexer;
+pub(crate) use actual_types::*;
+pub(crate) use indent_aware_lexer::IndentAwareYarnSpinnerLexer as YarnSpinnerLexer;
