+++ crates/yarn_slinger/tests/error_handling_tests.rs
@@ -1,7 +1,7 @@
 //! Adapted from <https://github.com/YarnSpinnerTool/YarnSpinner/blob/da39c7195107d8211f21c263e4084f773b84eaff/YarnSpinner.Tests/ErrorHandlingTests.cs>
 
 use crate::test_base::*;
-use yarn_slinger_compiler::prelude::*;
+use yarn_slinger::prelude::*;
 
 mod test_base;
