+++ crates/yarn_slinger/tests/type_tests.rs
@@ -15,8 +15,7 @@
 //! - `TestBuiltinTypesAreEnumerated`
 
 use crate::test_base::*;
-use yarn_slinger_compiler::prelude::*;
-use yarn_slinger_core::prelude::*;
+use yarn_slinger::prelude::*;
 
 mod test_base;
